Fix annoying bug in update_story_sentence_words.
update_story_sentence_words was fine for initially creating the word vectors for a story but it broke when trying to update the word vectors.

The problem was that since the story had already been updated once every sentence was considered a duplicate since every sentence would show up in the story_sentence_counts table. The fix is to delete sentences from this table if they came from the story being updated.