Fix subtle and nasty in update_aggregate_words
Explicitly convert the date to YYY-MM-DD format.

Make update_weekly_words work if the start_date is not the start of the week. Previously there was a subtle bug: words were deleted from the table by using postgresql 'datetrunc' week on the date param to get the publish week. Words were added to the table by looking at the date param and the next 6 days. This the date param wasn't the start of the week these two sets of dates were different.