Remove extract_for_crawler() and 'extract_in_process'
extract_for_crawler() was doing double-work. In normal crawler
operation, it was trying and retrying indefinitely to add a crawled
download to the extractor queue on the job broker. However, when running a
crawler unit test (and 'extract_in_process' configuration variable was
set -- not in mediawords.yml, but confusingly in the crawler unit test),
the subroutine was extracting the download directly in the crawler
process.

In this commit, ::Crawler::Engine gets a new 'extract_in_process'
parameter which, when seti (directly or by enabling crawler's
'test_mode'), will extract the download in extractor process instead of
adding it to the job broker's queue.

Also, crawler now doesn't retry adding an extraction job to the job
queue indefinitely (except for retrying to connect to RabbitMQ for a
minute like the rest of the code). If extractor's chain was broken (e.g.
due to RabbitMQ failure), crawler should be able to continue operation.

References #59.