restore command line prints and use public set membership and number of stories as sorting metrics
prints in this script got turned into logging message in a log4perl update.  These should actually be
prints (or says) because this is an interactive command line script.

After updating the dup_media_id in the database so that the parent medium is the one we want to search,
we need to update this deduping script to encourage choosing the best searched medium.  We do that
by sorting possible medium alternatives by public set membership and number of stories in the last
year.

Also removed inlink querying because we don't need now that we are looking at the number of stories and
because the inlink query is very slow.  Also optimized some of the queries so that the script only takes
a minute or two to get started.