add per topic permissions to api; add test_api_permissions to test permissions of all api end points
This commit:
* adds user level read/write/admin permissions to the topics api as documented in the topics api spec;
* adds test_api_permissions.t to test that new permissions system as well as the public / non-public key
permissions of the api;
* cleans up some existing permissions bugs surfaced by the testing;
* especially removes the Local flag from all func_METHOD (eg. list_GET) methods in the api to
disallow web access to those underyling functions.

test_api_permissions.t has a hard coded list of the permissions it expects for every end point. It
also queries catlyst directly for a list of all /api/* end points and fails a test of there are
any /api/* end points not tested by the script.