untested code to export timespans_id to solr
We are currently searching topics by using the psuedo query code to translate {~ .. } clauses
into long lists of stories_ids on the perl side of the api.  This worked fine when I wrote the code
for dealing with topics of ~10k stories, but we now have lots of topics with ~100k stories and the
giant election topic with > 1 million stories, so this is obviously a crappy solution.

This commit makes the solr import code query the timespans_id from postgres and index it in solr.  It
also makes TM/Snapshot.pm add all topic stories to solr_import_extra_stories so that they will get
reindexed.  We still use the search by stories_ids to generate foci within the snapshot, since we don't
want to wait for the soilr import to be able to generate the foci, but I have updated that code to do
the solr search for stories_ids in 10k chunks so that we don't run into solr query size limits even
for very big topics.  Performance for this approach is likely to be poor, but it should be good enough for
now (and is the only thing that doesn't require major re-architecture of our indexing process).