change default configure not to start any supervisord programs but to use the default rabbitmq configuration
travis seems to be starting supervisord for some reason on its own and then taking forever to do anything
because the machine is thrashing trying to get all of the supervisor programs running.  this changes makes
the default behavior on a fresh install not to run any supervisor programs via a setting in mediawords.yml.

this also enables the default rabbitmq config in mediawords.yml.dist so that it will work out of
the box without having to edit mediawords.yml.