change media_health generation to use existing tables so that it works during database dump
the mediawords_generate_media_health.pl script was dropping and recreating tables.  this does
not work when the database schema is locked during a database dump, so it breaks while the
database is dumping.  this commit changes the code to use delete + insert instead of drop + create.
it also moves the health generation code into MediaWords::DBI::Media::Health so that
we can write a test for it, and adds a simple test for media health generation in
t/test_media_health.t.