not yet working code to translate solr queries into postgres tsquery format
we currently have to hand write a regular expression for each topic, because we need to check whether
a given web page matches the topic without waiting to stick it into solr.  this code will eventually
be used to instead just convert the solr query directly to the postgres tsquery format so that we
can test whether a given block of text matches a solr query without first putting it into solr.

this will also eventually allow us to generate word clouds based on the specific sentences that
match a given solr query without having to store the sentences or text in solr.