change twitter topics to be a stand alone topics that run on their own
this update removes the code that automatically runs a child twitter topic for a parent web topic.  the
original idea was to make the child topic a seamless part of the parent topic, but in practice they have
been used as separate topics, and the integration of the two topics is super confusing for the
data architecture.  also, the twitter topics have worked quite well, and it is very likely we will
have more types of topics in the future, and making them cleanly separate is a step in that direction.

this change is in the sources_api branch because it also cleans up a bug caused by the nested topic
mining interacting with the new job_states stuff.