Increase magic sleep(1) in provide_downloads() to 5 seconds
Apparently provide_downloads() is waiting for the "engine" to finish
something and expects that it will do that in a single second; if
"engine" doesn't finish in said second, a given run to
provide_downloads() won't provide the engine with anything, thus making
the crawler test fail.

Python's UserAgent made a single download slightly slower (due to
Inline::Python interoperation probably), so a single download was being
fetched in >1 second.