On "requests" exceptions, set response's body to the exception message
As per LWP::UserAgent's (from which we inherit our UserAgent class
design) philosophy, if the parameters are right, UserAgent should never
raise and instead always return a valid Response object. To see whether
the request has failed, one is expected to call is_success() method.

On "requests" failures (when "requests" crashes being unable to parse
something), we create a Response object for with the exception message
in its body.

Before, this was being done incorrectly. In this commit, we set the body
correctly (in a way that can later be read by response.raw.stream()) and
also test whether the exception message got returned in HTTP body.