Rename get_stop_words() to stop_words_map(), remove stopword caching
Renamed to be more consistent with the rest of the names; also, "_map"
suffix suggests that the subroutine returns a hashref (which it does).

stop_words_map() now is to be implemented in every language subclass,
and then the subclass is to decide where to load the stopwords from.

Stopwords file caching was removed because it's superfast to load them
from a plain text file everytime; also, the single current user
(::Solr::WordCounts) calls stop_words_map() only once for every language
so the stopwords hashref caching was not being effective.