Don't use __repr__ incorrectly
If done my the book, __repr__ should return something that could be eval'ed back into that object.

We only need to be able to stringify the Request / Response object (mostly for debugging purposes) so maintaining an implementation of __repr__ is pointless.

Also, Python is probably making some brave assumptions about __repr__ so we get weird errors such as:

     SystemError: PyEval_EvalFrameEx returned a result with an error set