update query -> regex translation to support proximity clauses, {!} clauses, and inclusive regexes
These updates are needed for story index support.  With the story index,
we will need to use proximity clauses commonly to be able to do
effective sentence based queries, so we will need to support them in the
regex generator.  likewise, we will need to use {!complexphrase} phrase
queries.  Finally, we will need to generate sentence based word clouds
based on story query matches.  To do that, we will generate an inclusive
regex from the query (that just matches any term within the query) and
return the word cloud for all sentences in each matching story that
match the inclusive regex.