Add helpers for training, storing and loading word2vec vectors
train_word2vec_model() expects a sentence iterator object that it will use to fetch sentences for training the model, and model store object to be used for storing the raw serialized model data to a key-value store in the database.

load_word2vec_model() uses a model store object to load and return the raw serialized model data from a key-value store.