Don't deep_copy or wantarray return value in HandlerProxy if the caller is python
When a database handle was created in perl and passed to python, the hashes() and flat() methods were returning
the single value rather than a list in python when the query returned only one row.  This is different than the
behavior when calling those functions from a python created db handle, so it was causing intermittent havoc in
python functions called from perl with a perl db handle.  The cause was that those methods were calling wantarray
in perl to figure out whether to return a list or a ref to a list, which was causing the return type confusion.

You can reproduce this issue by uncommenting the new code in HandlerProxy hashes() and running
lib/MediaWords/DB/t/HandlerProxy.t

In the theory of reducing unnecessary work and round trip complexity, I added a check to all of the deep_copy(
$return_value ) calls in HandlerProxy.pm to avoid the deep_copy if the caller is python.