Install pkg-config for CyHunspell to be able to find Hunspell
CyHunspell, Hindi's stemming dependency, depends on Hunspell which it
tries to find using pkg-config. If it fails to find it, CyHunspell's
installer tries to download and compile Hunspell package from
SourceForge. downloads.sourceforge.org gets constantly down recently and
so some of the CI runs fail just because of that.