cleanup processing flow in Mine.pm; fix bug causing some topic_links to get lost
There was a bug that causing some stories_ids not to get assigned to the ref_stories_id value of the
source link_topics row, essentially making the link disappear after fetching.  To fix that bug, I had
to redesign some of the flow to save the topic_links_id in topic_fetch_urls and pass topic_fetch_urls
rows around in add_new_link_chunks() instead of stories.  In the process or rewriting to make that change
I moved much of functionality, including failed url checking and initial story matching, into the
fetch_link job.