try requeueing fetch_link jobs before timing out entirely
for some reason, topics are failing frequently because they are timing out waiting for one or two
fetch_link jobs to finish.  when I check the queue, those topic_fetch_url rows are sitting in the
database in pending or requeued state but seem to have been dropped from the rabbit queue.  this is
likely because the jobs are crashing for some reason and failing to update the topic_fetch_url state
to 'python_error', but I haven't been able to figure out how that's happening.  rerunning one of these
failed topics almost always makes it work the second time, so until we can find the underlying issue
with the fetch_link job or rabbit, I am going to just resubmit jobs that look like they have been
lost somehow.