optimize solr import query that pulls stories from queue table
when the queue table is large, the query that pulls stories_ids from it
eventually gets very slow because it has to work its way all of the dead
deleted rows.  this update makes that query keep track of the current
last max dequeued stories_id so that we can avoid hitting all of those
dead rows.