Add INSERT trigger on story_sentences_partitioned; change order of triggers so that last_updated_trigger() fires first
* A trigger on the intermediary `story_sentences_view_insert_update_delete` view was
INSERTing new sentences directly into the partitions and serving as a
INSERT trigger on `story_sentences_partitioned` itself. This behavior
was not firing `last_updated_trigger()` on
`story_sentences_partitioned` base table. The trigger on the view will
now INSERT directly into `story_sentences_partitioned` base table,
letting it choose which partition to insert to further.

* Added INSERT trigger on `story_sentences_partitioned` which didn't
exist before (because it didn't have to -- no one was supposed to use
`story_sentences_partitioned` directly, all the operations were to be
done on the `story_sentences` intermediary view).

* Changed order of `BEFORE` triggers on `story_sentences_partitioned` by
renaming them. When multiple triggers are to be fired for the specific
event on a table, PostgreSQL runs them in an alphabetical order.
`last_updated_trigger` (starting with an "l") had to fire before
`insert_trigger`, so they've been renamed to
`00_last_updated_trigger` and `01_insert_trigger`.