Create temporary table with snapshot's stories_id to make query planner avoid sequential scans
For whatever reason, fetching snapshot's stories_id in a subselect makes the query planner opt for a sequential scan on "story_sentences", despite the fact that planner's prediction on how many rows it expects to find in snap.stories is about right.

Creating a temporary table with snapshot's stories_id first and querying against that table seems to be a workaround against it.

Not using get_temporary_ids_table() because I'm not sure how many stories_id there could be; biggest snapshot at the time of writing is around ~1m sentences.

Fixes #348.