replace media_normalized_urls table with media.normalized_url
the existing media_normalized_urls table setup required the existence of
the media.db_row_last_updated field to work efficiently.  because we are
getting rid of that field, we need another way of keeping the media
normalized_urls up to date from python.  adding a normalized_url field
to the media table is the simplest way to do that, at the cost of
requiring the coder to null existing media.normalized_url values in the
database whenever updating the mediawords.tm.media.normalize_url_lossy
function.