Return tag search FTS conditional to _fetch_list() instead of temporary table
Some popular tag queries ('Trump') might generate hundreds of thousands
of tags_id which then have to be used as a condition by the main SELECT
in _fetch_list() only to return 20 or so rows (by default).

Instead of a temporary table, just return the FTS condition back to
_fetch_list() and trust the database to pick the most effective query
plan given the tags_id offset and LIMIT.

References #495.