optimize provider queries
now that we are using the recursive query to find one downloads_id per
host, it is easy to efficiently ignore the downloads_ids already in
queued_downloads, so we don't need to update the download state to
fetching.  this will save about 80% of the current time in the download
provider loop.