remove instant exit feature of topic mine fetch_link()
the topic mine fetch_link() routine fetches a chunk of 10k links by
submitting them to the pool of fetch_link workers and polling to check
when they are done fetching.  the limiting factor for the fetch_link
pool performance is per site throttling, so the existing code just
requeues all remaining links with a 0 second delay for throttling and
then exits, assuming that the small number of remaining urls will get
fetched quickly.  this works in practice and improves performance a bit,
but it makes tests fail occasionally on slower machines.  this commit
changes this functionality to requeue the remaining links with the 0
throttle but still wait around to make sure they are collected.