complete by untested support for archive.org topic_tweets
this commit also adds the topic_seed_queries table to add additional
seed queries for each topic, and the concept of a platform (eg. twitter)
and a source (eg. crimson_hexagon or archive_org) for each seed query as
well as a platform for the topic to distinguish web topics from twitter
topics.

once tested the code should work as a stand in replacement, requiring
only a topic_seed_query with a archive_org source to fetch the data from
archive.org.