add efficient incremental respidering support to topics
this commit adds two changes.  it changes the topics api to simply mark
a topic as needing story respidering when the query or date range
changes, then changes the topic spider to actually update the
topics_stories rows to mark them for respidering.  it also updates that
respidering logic to move the timespans of a snapshot that is being thus
respidered to archive any timespans within any date changes, so that
only impacted timespans will be regenerated but no timespan data will
actually be deleted.