In dev/test environments, mount source code of dependent services too
Before, if one runs a test in an app A (e.g. "topics-mine"), and the test depends on source code changes in dependency app B (e.g. "extract-and-vector"), one has to rebuild dependency app B's container image for its source code changes to be reflected while running a test in A.

After, source code gets mounted to both the tested app A's container and all of its dependencies, so source code changes in both A and its dependencies get reflected without having to rebuild neither A's nor dependency B's container image.