remove transactions from extraction
the transactions in the extract_and_vector jobs are hanging other jobs
because the extractor jobs die somehow and end up stuck in the
transaction.  these transactions don't add much.  the cost is that we
will have downloads in the extracted=true state that still need to be
processed for story_sentences, but we should be looking at
processed_stories to figure out whether a story has been processed in
any case.