update facebook error handling
further update facebook error handling so that it requires more than
three 500s with a blank response content to kill the worker.  otherwise
just assume that there is something in the url that is causing facebook
to 500.  also add a test for the facebook error handling, which has
enough logic now to warrant it.