change topic mining jobs to extract jobs in place
To avoid an extract_and_vector dependency for topics, we are
sending extraction jobs from topic fetch_link jobs to the extractor
work pool.  This is majorly slowing down the link fetching as we wait
for the extractor pool to get around to handling he jobs.  We could
implement priority within the extract_and_vector queue, but that is
still going to be a less efficient than just extracting jobs
in process.  The stability and performance of the fetch_link jobs is
critical, so it is worth having the extra dependency just to keep
this code path simple and fast.