Revert "shorten error message for run.py subprocess error"
`run.py` runs the following commands in sequence:

1) `docker-compose run <service_name> <command>`
2) `docker-compose down`

The first command might sometimes fail, e.g. if the command is running a
test. If the command fails, `run.py` still must run `docker-compose
down` in order to shut down all of the services in the Compose
configuration and clean up all the volumes; otherwise, subsequent runs
of the same `run.py` command might lead to it reusing the previously
created containers and volumes; in case of a unit test being run in such
a setup, it might lead to unexpected results.

This is why `subprocess.call()` + printing an error message instead of
`check_call()` doesn't cut it - `call()` doesn't raise an exception that
could be caught and remembered by `run.py` to be later thrown to
`run.py` script's caller.

This reverts commit 40c1350cd2a26d3fea907f48d9f722eaa38e330b.