refuse to refetch fb data for story with error
a few times now, a single story that consistently causes a weird error
has killed our whole facebook fetching queue because it just sits at the
top of the queue killing every worker that picks it up.  this change
makes the fetcher refuse to try to fetch a story that already has an
error to prevent this from happening.  if we want to refetch a given
story, we can just set its error in story_statistics to null.