Set idle_transaction_timeout to 10 minutes in PgBouncer
If the client application is misconfigured / buggy and leaves the transaction open for a long period of time, the connection then ends up in an "idle in transaction" state, the snapshot before the transaction started has to be kept around indefinitely and thus the autovacuum breaks, making the database run at a risk of hitting the transaction ID wraparound.

With this in place, PgBouncer is supposed to kill off "idle in transaction" connections after 10 minutes, i.e. if the client doesn't get around to executing COMMIT / ROLLBACK in 10 minutes, it will get disconnected.